Rename snake_case source files to PascalCase in src/config
Update all #include directives and CMakeLists.txt source paths
to match the new filenames.
