+++ src/config/Loader.cpp
@@ -1,4 +1,4 @@
-#include "config/loader/loader.h"
+#include "config/loader/Loader.h"
 
 #include <stdexcept>
