+++ test/config/loader_test.cpp
@@ -1,4 +1,4 @@
-#include "config/loader/loader.h"
+#include "config/loader/Loader.h"
 
 #include <gmock/gmock.h>
 #include <gtest/gtest.h>
